Split uploader from formatter
