+++ codecov.gemspec
@@ -8,7 +8,7 @@ Gem::Specification.new do |s|
   s.summary               = 'Hosted code coverage'
   s.description           = 'Hosted code coverage Ruby reporter.'
   s.email                 = ['hello@codecov.io']
-  s.files                 = ['lib/codecov.rb', 'lib/codecov/version.rb']
+  s.files                 = ['lib/codecov.rb', 'lib/codecov/formatter.rb', 'lib/codecov/uploader', 'lib/codecov/version.rb']
   s.homepage              = 'https://github.com/codecov/codecov-ruby'
   s.license               = 'MIT'
   s.platform              = Gem::Platform::RUBY
